extract table in its own TableView component so we have a bit cleaner code
